+++ RELEASE-NOTES.md
@@ -1,3 +1,7 @@
+### Version 2.3.0 - Feb 19, 2018
+
+- **Accordion** - Added `onChanging` callback for accordion that occurs before animation in both directions **Thanks @GammeGames** [#5892](https://github.com/Semantic-Org/Semantic-UI/pull/5892)
+
 #### Bugs
 
 - **Accordion** - Added missing notation for accordion docs [#2812](https://github.com/Semantic-Org/Semantic-UI/issues/2812)